+++ quiche/src/lib.rs
@@ -17802,7 +17802,7 @@ pub use crate::path::PathEvent;
 pub use crate::path::PathStats;
 pub use crate::path::SocketAddrIter;
 
-pub use crate::recovery::congestion::CongestionControlAlgorithm;
+pub use crate::recovery::CongestionControlAlgorithm;
 
 pub use crate::stream::StreamIter;
